% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_humidity_data.R
\name{dewpoint_temperature}
\alias{dewpoint_temperature}
\title{Dew Point Temperature.}
\usage{
dewpoint_temperature(tm, rh)
}
\arguments{
\item{tm}{the air temperature in degree Celsius.}

\item{rh}{the relative humidity in percentage.}
}
\value{
The dew point temperature in degree Celsius.
}
\description{
Compute the dew point temperature.
}
