% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_metar_messages.R
\name{get_Metar_Data}
\alias{get_Metar_Data}
\title{Parse METAR code.}
\usage{
get_Metar_Data(metar)
}
\arguments{
\item{metar}{character, the full strings of the metar code.\cr
Example: "METAR HUEN 230700Z 08008KT 9999 SCT022 FEW025 26/22 Q1016 NOSIG="}
}
\value{
a list of the ICAO location indicator, day, hour, minute, wind speed and direction, 
temperature, dew point and QNH.
}
\description{
Get data from METAR strings of weather reports.
}
\examples{

\dontrun{
library(decodeSYNOP)

metar <- "METAR HUEN 230700Z 08008KT 9999 SCT022 FEW025 26/22 Q1016 NOSIG="
out <- get_Metar_Data(metar)
}

}
