% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_some_variables.R
\name{get_Synop_Data}
\alias{get_Synop_Data}
\title{Surface observation report from land station (FM-12).}
\usage{
get_Synop_Data(synop)
}
\arguments{
\item{synop}{a synop data object, output of the function \code{getSynopSections}.}
}
\value{
A data frame of one row, the columns contain the variables.
}
\description{
Get the data from a selected variables from a SYNOP (AAXX) strings of weather reports.
}
\examples{

\dontrun{
aaxx <- "AAXX 07181 33837 11583 83102 10039 21007 30049 40101 52035 60012 70282 8255/="
synop <- getSynopSections(aaxx)
df <- get_Synop_Data(synop)
}

}
