% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_actuals_messages.R
\name{get_Vekn_Data}
\alias{get_Vekn_Data}
\title{Parse ACTUALS messages.}
\usage{
get_Vekn_Data(vekn)
}
\arguments{
\item{vekn}{character, the full strings of the \code{VEKN} code.\cr
Example: "VEKN20 HKNC 260600 63741 MAX 26.3 MIN 17.7 A/R NIL RH 81\%="}
}
\value{
a list of the ICAO ID, WMO ID , day, hour, minute, precipitation,  
maximum and minimum temperature, dry and wet bulb, dew point and vapor pressure.
}
\description{
Get data from ACTUALS VEKN strings.
}
\examples{

\dontrun{
library(decodeSYNOP)

vekn <- "VEKN20 HKNC 260600 63741 MAX 26.3 MIN 17.7 A/R NIL RH 81\%="
out <- get_Vekn_Data(vekn)
}
## different forms of vekn messages
# "VEKN20 HKNC 281200 63740 DBT=28.0 WBT=18.9 VP=15.9 RH=42\%="
# "VEKN72 HKGA 250600 63619 MAX-28.0 MIN-19.5 A/R-NIL R/H-85\%="
# "VEKN71 HKJK 250600 63737 MAX: 27.9; MIN: 13.5; A/R: 0.0MM; R/H: 67\%="
# "VEKN72 HKGA 250600 63723 TX 34.9, TN 25.7, A/RF TR,R.H 75\%="
# "VEKN73 HKKI 250600 HKKR(63710) MX=27.5,MN=8.5,R/H=74\%,A/RF=NIL="
# "VEKN75 HKNC 250600 63746 MAX27.9 MIN18.0 GMIN15.1 ARF1.8 RH87\%="
# "VEKN73 HKKI 290600 63686 MAX.22.0 MIN.11.1 A/RF.0.0MM RH.60\%="
# "VEKN75 HKNC 280600 KANGEMA RANET ARFNILL DB20.1 DB16.2 MAX24.8 MIN13.7 DP14.0 RH67\% CLD1CI VV30KM WIND24002 GM13.0="

}
