% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temperature_data.R
\name{getAirTemperature}
\alias{getAirTemperature}
\title{Get Air Temperature.}
\usage{
getAirTemperature(synop)
}
\arguments{
\item{synop}{a synop data object, output of the function \code{getSynopSections}.}
}
\value{
The air temperature observation in degree Celsius.
}
\description{
Get the air temperature observation, group \code{1s_nTTT} from section 1.
}
\examples{

\dontrun{
aaxx <- "AAXX 28154 89022 42698 72506 11360 21382 39758 49808="
synop <- getSynopSections(aaxx)
tmp <- getAirTemperature(synop)
}

}
