% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_humidity_data.R
\name{getDewPointTemperature}
\alias{getDewPointTemperature}
\title{Get Dew Point Temperature.}
\usage{
getDewPointTemperature(synop)
}
\arguments{
\item{synop}{a synop data object, output of the function \code{getSynopSections}.}
}
\value{
The dew point temperature in degree Celsius.
}
\description{
Get the dew point temperature, group \code{2s_nT_dT_dT_d} from section 1.
}
\examples{

\dontrun{
aaxx <- "AAXX 07181 33837 11583 83102 10039 21007 30049 40101 52035 60012 70282 8255/="
synop <- getSynopSections(aaxx)
td <- getDewPointTemperature(synop)
}

}
