% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temperature_data.R
\name{getMaximumTemperature}
\alias{getMaximumTemperature}
\title{Get Maximum Temperature.}
\usage{
getMaximumTemperature(synop)
}
\arguments{
\item{synop}{a synop data object, output of the function \code{getSynopSections}.}
}
\value{
The maximum temperature observation in degree Celsius.
}
\description{
Get the maximum temperature observation, group \code{1s_nT_xT_xT_x} from section 333.
}
\examples{

\dontrun{
aaxx <- "AAXX 28154 89022 42698 72506 333 10312 20195="
synop <- getSynopSections(aaxx)
tmax <- getMaximumTemperature(synop)
}

}
