% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temperature_data.R
\name{getMinimumTemperature}
\alias{getMinimumTemperature}
\title{Get Minimum Temperature.}
\usage{
getMinimumTemperature(synop)
}
\arguments{
\item{synop}{a synop data object, output of the function \code{get_sectionSynop}.}
}
\value{
The minimum temperature observation in degree Celsius.
}
\description{
Get the minimum temperature observation, group \code{2s_nT_nT_nT_n} from section 333.
}
\examples{

\dontrun{
aaxx <- "AAXX 28154 89022 42698 72506 333 10312 20195="
synop <- getSynopSections(aaxx)
tmin <- getMinimumTemperature(synop)
}

}
