% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_datetime.R
\name{getObsActualTime}
\alias{getObsActualTime}
\title{Get the actual time of observation.}
\usage{
getObsActualTime(synop)
}
\arguments{
\item{synop}{a synop data object, output of the function \code{getSynopSections}.}
}
\value{
A character vector of length 2, c(hour, minute). If the \code{9GGgg} group does not exist it returns the hour from the \code{YYGGiw} group and the minute is set to "00".
}
\description{
Get the actual time of observation if exist, group \code{9GGgg} of section 1.
}
