% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_datetime.R
\name{getObsDateTime}
\alias{getObsDateTime}
\title{Get Observation Date-Time.}
\usage{
getObsDateTime(synop)
}
\arguments{
\item{synop}{a synop data object, output of the function \code{getSynopSections}.}
}
\value{
A character vector of length 2, c(day, hour).
}
\description{
Get the day and hour of a meteorological report, group \code{YYGGiw} of section 1.
}
\examples{

\dontrun{
aaxx <- "AAXX 28154 89022 42698 72506 11360 21382 39758 49808="
synop <- getSynopSections(aaxx)
ddhh <- getObsDateTime(synop)
}

}
