% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_humidity_data.R
\name{getRelativeHumidity}
\alias{getRelativeHumidity}
\title{Get Relative Humidity.}
\usage{
getRelativeHumidity(synop)
}
\arguments{
\item{synop}{a synop data object, output of the function \code{get_sectionSynop}.}
}
\value{
The relative humidity in percentage.
}
\description{
Get the relative humidity, group \code{29UUU} from section 1.
}
\examples{

\dontrun{
aaxx <- "AAXX 07181 33837 11583 83102 10039 29072 30049 40101 52035 60012="
synop <- getSynopSections(aaxx)
rh <- getRelativeHumidity(synop)
}

}
