% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pressure_data.R
\name{getSeaLevelPressure}
\alias{getSeaLevelPressure}
\title{Get Sea Level Pressure.}
\usage{
getSeaLevelPressure(synop)
}
\arguments{
\item{synop}{a synop data object, output of the function \code{getSynopSections}.}
}
\value{
The sea level pressure in hectopascal.
}
\description{
Get the sea level pressure, group \code{4PPPP} from section 1.
}
\examples{

\dontrun{
aaxx <- "AAXX 28154 89022 42698 72506 11360 21382 39758 49808="
synop <- getSynopSections(aaxx)
slp <- getSeaLevelPressure(synop)
}

}
