% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_datetime.R
\name{getStationID}
\alias{getStationID}
\title{Get WMO station ID.}
\usage{
getStationID(synop)
}
\arguments{
\item{synop}{a synop data object, output of the function \code{getSynopSections}.}
}
\value{
A character, the WMO station ID.
}
\description{
Get WMO station ID, group \code{IIiii} of section 1.
}
\examples{

\dontrun{
aaxx <- "AAXX 28154 89022 42698 72506 11360 21382 39758 49808="
synop <- getSynopSections(aaxx)
stnID <- getStationID(synop)
}

}
