% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pressure_data.R
\name{getStationPressure}
\alias{getStationPressure}
\title{Get Station Pressure.}
\usage{
getStationPressure(synop)
}
\arguments{
\item{synop}{a synop data object, output of the function \code{get_sectionSynop}.}
}
\value{
The station pressure in hectopascal.
}
\description{
Get the station pressure, group \code{3P_oP_oP_oP_o} from section 1.
}
\examples{

\dontrun{
aaxx <- "AAXX 28154 89022 42698 72506 11360 21382 39758 49808="
synop <- getSynopSections(aaxx)
pres <- getStationPressure(synop)
}

}
