% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_synop_sections.R
\name{getSynopSections}
\alias{getSynopSections}
\title{Parse SYNOP code.}
\usage{
getSynopSections(synop)
}
\arguments{
\item{synop}{character, the full strings of the synop code.\cr
Example: "AAXX 28154 89022 42698 72506 11360 21382 39758 49808="}
}
\value{
A synop data objects. It is a list object with elements the sections [1, 222//, 333, 555] of the synop.
}
\description{
Parse SYNOP (AAXX) strings of weather reports.
}
\examples{

\dontrun{
library(decodeSYNOP)

aaxx <- "AAXX 28154 89022 42698 72506 11360 21382 39758 49808="
synop <- getSynopSections(aaxx)
}

}
