% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wind_data.R
\name{getWindData}
\alias{getWindData}
\title{Get wind measurement.}
\usage{
getWindData(synop)
}
\arguments{
\item{synop}{a synop data object, output of the function \code{getSynopSections}.}
}
\value{
A numeric vector of length 2, c(direction [in degree (0-360)], speed [in m/s]).
}
\description{
Get wind data, group \code{Nddff} and \code{00fff} if exist.
}
\examples{

\dontrun{
aaxx <- "AAXX 28154 89022 42698 72506 11360 21382 39758 49808 52006 875// 333 87635="
synop <- getSynopSections(aaxx)
wnd <- getWindData(synop)
}

}
