% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_GTS_HKNC.R
\name{parse_GTS_Bulletin}
\alias{parse_GTS_Bulletin}
\title{Parse GTS Bulletins.}
\usage{
parse_GTS_Bulletin(bulletin_file)
}
\arguments{
\item{bulletin_file}{full path to the file containing the GTS bulletin.}
}
\value{
A vector of the SYNOP (AAXX) strings of weather reports.
}
\description{
Parse GTS bulletin files from Regional Telecommunication Hub (RTH).
}
\examples{

\dontrun{
# GTS bulletin file from Kenya Meteorological Department (HKNC)
file_HKNC <- "~/HKNC_bulletins/HKNC50044537.a"
synop_strings <- parse_GTS_Bulletin(file_HKNC)
}

}
