% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_GTS_HKNC.R
\name{parse_GTS_HKNC_Bulletins}
\alias{parse_GTS_HKNC_Bulletins}
\title{Parse GTS HKNC Bulletins.}
\usage{
parse_GTS_HKNC_Bulletins(file_HKNC)
}
\arguments{
\item{file_HKNC}{full path to the file containing the GTS bulletins.}
}
\value{
A vector of the SYNOP (AAXX) strings of weather reports.
}
\description{
Parse GTS bulletins file from Kenya Meteorological Department (HKNC).
}
\examples{

\dontrun{
file_HKNC <- "~/HKNC_bulletins/HKNC50044537.a"
synop_strings <- parse_GTS_HKNC_Bulletins(file_HKNC)
}

}
