% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_humidity_data.R
\name{relative_humidity}
\alias{relative_humidity}
\title{Relative humidity.}
\usage{
relative_humidity(tm, td)
}
\arguments{
\item{tm}{the air temperature in degree Celsius.}

\item{td}{the dew point temperature in degree Celsius.}
}
\value{
The relative humidity in percentage.
}
\description{
Compute the relative humidity.
}
