% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_humidity_data.R
\name{saturation_vapour_pressure}
\alias{saturation_vapour_pressure}
\title{Saturation Vapour Pressure.}
\usage{
saturation_vapour_pressure(tmp)
}
\arguments{
\item{tmp}{the air temperature in degree Celsius.}
}
\value{
The saturation vapour pressure in hectopascal.
}
\description{
Compute the saturation vapour pressure.
}
