% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_humidity_data.R
\name{specific_humidity}
\alias{specific_humidity}
\title{Specific humidity.}
\usage{
specific_humidity(td, pr)
}
\arguments{
\item{td}{the dew point temperature in degree Celsius.}

\item{pr}{the surface pressure in hectopascal.}
}
\value{
The specific humidity in kg/kg.
}
\description{
Compute the specific humidity.
}
