% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_precipititation_data.R
\name{getPrecipitation}
\alias{getPrecipitation}
\title{Get precipitation data.}
\usage{
getPrecipitation(synop)
}
\arguments{
\item{synop}{a synop data object, output of the function \code{get_sectionSynop}.}
}
\value{
A named list of numeric vector of length 2, c(duration[in hour], amount [in mm]).\cr 
If \code{i_R} from the group \code{i_Ri_XhVV} of section 1 equals 0, it returns a list of length 2,\cr
first element is the precipitation from section 1 group \code{6RRRt_R} \cr
and the second element from section 333 group \code{7R_24R_24R_24R_24}.
}
\description{
Get precipitation data, group \code{6RRRt_R} from section 1 and group \code{7R_24R_24R_24R_24} from section 333.
}
\examples{

\dontrun{
aaxx <- "AAXX 07181 33837 11583 83102 10039 29072 30049 40101 52035 60012="
synop <- getSynopSections(aaxx)
pres <- getPrecipitation(synop)
}

}
